basic keyboard implementation, changes to the entities, some test entities

+++ Engine/Components/Components.h
@@ -38,10 +38,10 @@ struct SpriteComponent : public Component {
 };
 
 struct VelocityComponent : public Component {
-    uint16_t v_x = 0;
-    uint16_t v_y = 0;
+    int8_t v_x = 0;
+    int8_t v_y = 0;
 
-    VelocityComponent(uint8_t x, uint8_t y) : v_x(x), v_y(y) {};
+    VelocityComponent(int8_t x, int8_t y) : v_x(x), v_y(y) {};
     VelocityComponent() {};
 };
